doc-gen(mergeDecoratorDocs): capture all the metadata docs
Iterate through the docs and merge all the of metadata docs
for each decorator doc into the decorator doc and remove it
so that it is not rendered in a page of its own
